compilation on other platforms

+++ ex/Test/test_highterrain.cpp
@@ -20,7 +20,7 @@ int main(int argc, char** argv)
   plan_tests(NUM_TERRAIN);
 
   for (int j=0; j<NUM_TERRAIN; j++) {
-    terrain_height = (800)*(j+1)/(NUM_TERRAIN*1.0);
+    terrain_height = (int)((800)*(j+1)/(NUM_TERRAIN*1.0));
 
     unsigned i = rand()%NUM_WIND;
     ok (test_flight_times(7,i), test_name("high terrain",7,i),0);
